+++ components/tricheckbox.tsx
